Simplify file structure
